fix: remove redundant tests
